pyproject.toml
[project]
name = "26fw-finml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "pandas>=3.0.1",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.0",
]
